+++ site/src/pages/chapters.js
@@ -1,10 +1,20 @@
 import React from "react";
 import ChaptersPage from "../sections/Chapters";
+import Layout from "../components/layout";
+import SeoComponent from "../components/seo";
+import Footer from "../components/Footer/index";
 
 const Chapters = () => {
   return (
     <>
-      <ChaptersPage />
+      <SeoComponent
+        title="Service Mesh Patterns - Chapters and Topics"
+        description="Chapters and topics found in the Service Mesh Patterns book"
+      />
+      <Layout>
+        <ChaptersPage />
+        <Footer />
+      </Layout>
     </>
   );
 };
